ARROW-1519: [C++] Move DecimalUtil functions to methods on the Int128 class
This is mostly moving things around so that there's one place to access all Int128 functionality from third parties, the original impetus being `parquet-cpp` decimal reads/writes.

Author: Phillip Cloud <cpcloud@gmail.com>
Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #1083 from cpcloud/ARROW-1519 and squashes the following commits:

27ae41dc [Wes McKinney] Rename Int128 to Decimal128
dd401c6d [Phillip Cloud] IWYU
51a3f240 [Phillip Cloud] Use a bool for negativity
94b5b37a [Phillip Cloud] Doc for Int128::FromString
db4d2070 [Phillip Cloud] Add doc and const
c06ebe83 [Phillip Cloud] Add DCHECK for Int128 pointer in StringToInteger
bec1c8b9 [Phillip Cloud] Add DCHECK for Int128 initial value in StringToInteger
d9829c9b [Phillip Cloud] Refactor Int128 construction from string
d7c2068c [Phillip Cloud] ARROW-1519: [C++] Move DecimalUtil functions to methods on the Int128 class

+++ cpp/src/arrow/builder.cc
@@ -35,7 +35,6 @@
 #include "arrow/util/cpu-info.h"
 #include "arrow/util/decimal.h"
 #include "arrow/util/hash-util.h"
-#include "arrow/util/int128.h"
 #include "arrow/util/logging.h"
 
 namespace arrow {
@@ -1119,7 +1118,7 @@ DecimalBuilder::DecimalBuilder(MemoryPool* pool, const std::shared_ptr<DataType>
     : DecimalBuilder(type, pool) {}
 #endif
 
-Status DecimalBuilder::Append(const Int128& value) {
+Status DecimalBuilder::Append(const Decimal128& value) {
   RETURN_NOT_OK(FixedSizeBinaryBuilder::Reserve(1));
   std::array<uint8_t, 16> bytes;
   RETURN_NOT_OK(value.ToBytes(&bytes));
